Extracted `fn cast_ptr_to_usize` to `crate::cast`.
